Move channel setup to a separate startNotify helper
Also add a very small buffer to the NotifyReturn channel
in case no one is listening on the other side.

I event thought about actually having a timeout when sending
to that channel, but it felt it could lead to loss of
events on non problematic situations, so let's see how this
works instead.