fix: goroutine leak in publisher on close
- closing publisher did not close all associated goroutines. The
routine listening for block events of connection only got terminated
on closing the connection.
The wrapping connection manager holds all blocking channels and only
passes down an universal blocking channel now. If a signal reaches this
channel it is broadcasted to all blocking channels. This allows telling
the connection manager to remove and close the channel if publisher is
closed.

Closes #149